pyproject.toml
[tool.poetry]
name = "donationportal_project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.4"
Pillow = "^9.3.0"
django-crispy-forms = "^1.14.0"
django-import-export = "^3.0.1"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.0.1"}
django-allauth = "^0.51.0"
django-filter = "^22.1"
social-auth-app-django = "^5.0.0"
django-environ = "^0.9.0"
django-storages = {extras = ["azure"], version = "^1.13.2"}
psycopg2-binary = "^2.9.5"

[tool.poetry.dev-dependencies]
Django = "^4.1.4"
Pillow = "^9.3.0"
django-crispy-forms = "^1.14.0"
crispy-bootstrap5 = "^0.7"
django-import-export = "^3.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
